Take common functions out of analysis directory
