+++ components/run_analysis/Runoff_Volume.py
@@ -13,6 +13,7 @@ import inspect
 
 cmd_folder = os.path.split(inspect.getfile(inspect.currentframe()))[0]
 sys.path.append(cmd_folder)
+sys.path.append(os.path.dirname(cmd_folder))
 
 
 from qnspect_utils import perform_raster_math
